Conflicts:
	Attrition.rc
	Attrition.vcxproj
	Attrition.vcxproj.filters
	bin/Settings/Graphics.lua
	include/Overlay.h
	include/State.h
	include/TestState.h
	include/modelObject.h
	src/Graphics.cpp
	src/LoggedGraphics.cpp
	src/LuaInterface.cpp
	src/TestState.cpp
	src/luaGraphics.cpp
	src/luaInterfaceAttrition.cpp
	src/main.cpp

+++ include/LuaState.h
@@ -21,8 +21,8 @@ namespace de
 				virtual ~LuaState();
 
 				bool handleEvents( const SDL_Event &_event );
-				bool logic( const Uint32 &_deltaTicks, State* &_nextState, state::options &_options );
-				bool logic( const Uint32 &_deltaTicks );
+				bool logic( const float &_deltaTicks, State* &_nextState, state::options &_options );
+				bool logic( const float &_deltaTicks );
 				void render();
 
 			private:
